Move some of the code around.
